Revert "move exported helpers to _common.go file"
This reverts commit eb54ad9a98dcd27da23331d41910a5c738edb303.
